Fixed the layout break of the HTML on conditional show
The prior script block that were moved up into the field section caused some problems in the layout when the captcha is hidden via condition. Since Caldera Forms moves the HTML into script tags of type "text/html" it caused to close the script tag too early by our own script tag, thus the layout after a hidden captcha would break. Now the bypass is to move the script block outside of the field into a function and call this function without using script tags. The only way to call a function without using script tags is to load a transparent small image. It's not beautiful, but the logic needs to trigger some javascript when the captcha is shown *every time* and again, script tags can't be used or the HTML would break.